+++ .rubocop_todo.yml
@@ -40,6 +40,7 @@ Lint/AssignmentInCondition:
   Exclude:
     - 'app/controllers/api/traces_controller.rb'
     - 'app/controllers/api/user_preferences_controller.rb'
+    - 'app/controllers/accounts_controller.rb'
     - 'app/controllers/application_controller.rb'
     - 'app/controllers/geocoder_controller.rb'
     - 'app/controllers/notes_controller.rb'
