+++ app/views/users/show.html.erb
@@ -28,7 +28,7 @@
               <%= link_to t(".my comments"), diary_comments_path(current_user) %>
             </li>
             <li>
-              <%= link_to t(".my settings"), user_account_path(current_user) %>
+              <%= link_to t(".my settings"), edit_account_path %>
             </li>
 
             <% if current_user.blocks.exists? %>
